fix: input left and right elements in pseudo props
